// src/app/enums/tour-steps.enum.ts
export const enum TourStepAnchorEnum {
  HEADER_LANGUAGE = 'headerLanguage',
  HEADER_SETTINGS = 'headerSettings',
  SETTINGS_TYPE = 'settingsType',
  SETTINGS_SIZE_TYPE = 'settingsSizeType',
  SETTINGS_SIZE = 'settingsSize',
  SETTINGS_BREAKER_QUANTITY = 'settingsBreakerQuantity',
  SETTINGS_PRINT_ORIENTATION = 'settingsPrintOrientation',
  SETTINGS_SEPERATED_UNITED = 'settingsSeparatedUnited',
  SETTINGS_LABEL_SIZE = 'settingsLabelSize',
  SETTINGS_FONT = 'settingsFont',
  SETTINGS_SAVE = 'settingsSave',
  CONSTRUCTOR_RAIL = 'constructorRail',
  ROW_EDIT_BREAKER = 'rowEditBreaker',
  EDIT_BREAKER_DIALOG_INFO = 'editBreakerDialogInfo',
  EDIT_BREAKER_DIALOG_COLOR = 'editBreakerDialogColor',
  EDIT_BREAKER_DIALOG_ICON = 'editBreakerDialogIcon',
  EDIT_BREAKER_DIALOG_ACTIONS = 'editBreakerDialogActions',
  ROW_DELETE_BREAKER = 'rowDeleteBreaker',
  CONSTRUCTOR_DELETE_ROW = 'constructorDeleteRow',
  CONSTRUCTOR_ADD_RAIL = 'constructorAddRail',
  HEADER_PREVIEW = 'headerPreview',
  PREVIEW_DIALOG_CONTENT = 'previewContent',
  PREVIEW_DIALOG_CLOSE = 'previewDialogClose',
  HEADER_DOWNLOAD = 'headerDownload',
}